pyproject.toml
[tool.poetry]
name = "synthetic medical record generator"
version = "0.1.1"
description = "SMRG is a practical and user-friendly tool designed to help create realistic synthetic medical records from ICD-10-CM and Current Procedural Terminology codes."
authors = ["jilp <leptonneutrino@protonmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
datasets = "^2.17.1"
openai = "^1.12.0"
python-dotenv = "^1.0.1"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"
numpy = "^1.26.4"
tenacity = "^8.2.3"
pytest = "^8.0.1"

# Black formatting

[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
flake8 = "^7.0.0"
isort = "^5.13.2"

[tool.black]
line-length = 150
include = '\.pyi?$'
exclude = '''
/(
      .eggs         # exclude a few common directories in the
    | .git          # root of the project
    | .hg
    | .mypy_cache
    | .tox
    | venv
    | _build
    | buck-out
    | build
    | dist
  )/
'''

# iSort
[tool.isort]
profile = "black"
line_length = 150
multi_line_output = 3
include_trailing_comma = true
virtual_env = "venv"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
